requirements.txt
discord.py
python-dotenv
asyncpg
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
youtube_dl
PyNaCl